% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temp_cnst_plot.R
\name{air.temp.cnst.plot}
\alias{air.temp.cnst.plot}
\title{Generate Pairwise Comparison Plots for SAAT Measurements by ML}
\usage{
air.temp.cnst.plot(site, bgn.month, end.month, save.dir)
}
\arguments{
\item{site}{The 4-letter NEON site code.}

\item{bgn.month}{The fisrt month of data to use in plotting.}

\item{end.month}{The last month of data to use in plotting.}

\item{save.dir}{The save directory for the output plot.}
}
\value{
A PNG of level-by-level comparison plots in the specified \code{save.dir}.
}
\description{
Produces  PNG of air temperature consistency plots, in pairwise measurement levels.
Plots are of temperature readings from ML-n versus readings from ML-n+1. The raw data used in the
plot are not saved.
}
\examples{
\dontrun{
air.temp.cnst.plot(site="CPER", bgn.month="2017-05", end.month="2017-05", save.dir=getwd())
}

}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{air}
\keyword{commissioning,}
\keyword{consistency,}
\keyword{data}
\keyword{process}
\keyword{quality,}
\keyword{temperature}
