% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temp_dq_test.R
\name{air.temp.dq.test}
\alias{air.temp.dq.test}
\title{Test NEON Air Temperature data for Stability of Variance, Internal Consistency,
and External Consistency.}
\usage{
air.temp.dq.test(site, bgn.month, end.month, save.dir)
}
\arguments{
\item{\code{site}}{Parameter of class character. The NEON site data should be downloaded for.}

\item{\code{dpID}}{Parameter of class character. The data product code in question. See
\code{Noble::tis_pri_vars} for a selected list of data product names and codes, or
\url{http://data.neonscience.org/data-product-catalog} for a complete list.}

\item{\code{bgn.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{\code{end.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{\code{time.agr}}{Parameter of class numeric. The data agregation interval requested, must be 1, 2, 5, or 30.}

\item{\code{package}}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{\code{save.dir}}{Parameter of class character. The local directory where data files should be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates and site, the function will perform variance stability testing
 on overnight data (midnight to 4:00 local time), internal consistency checks on adjacent
 measurement locations, and external consistency checks with the closest NRCS or USCRN site.
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{"2017-04",}
\keyword{"2017-05",}
\keyword{"CPER",}
\keyword{"DP1.00002.001",}
\keyword{#Make}
\keyword{30,}
\keyword{=}
\keyword{a}
\keyword{bgn.month}
\keyword{commissioning}
\keyword{data}
\keyword{data.pull(site}
\keyword{direcotry}
\keyword{dpID}
\keyword{end.month}
\keyword{example:}
\keyword{for}
\keyword{gaps,}
\keyword{package="basic",}
\keyword{process}
\keyword{quality,}
\keyword{save.dir=}
\keyword{tempDir)}
\keyword{tempDir<-}
\keyword{tempdir()}
\keyword{temporary}
\keyword{the}
\keyword{time.agr}
