% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temp_plot.R
\name{air.temp.plot}
\alias{air.temp.plot}
\title{Air Temperature Plot}
\usage{
air.temp.plot(bgn.month, end.month, site, save.dir)
}
\arguments{
\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to plot.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to plot.}

\item{site}{The NEON TIS site to generate plots for.}

\item{save.dir}{Directory where plots should be written.}
}
\value{
A data frame with data collected between \code{bgn.time} and \code{end.time}.
}
\description{
Plots and saves all temperature measurements at a site for a given date range.
}
\examples{
\dontrun{
save.dir=getwd()
bgn.month<-"2016-01"
end.month<-"2016-04"
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{data,}
\keyword{date,}
\keyword{subset,}
