% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ais_site_config}
\alias{ais_site_config}
\title{Site Configuration for NEON AIS Stations}
\format{Information on all 36 AIS sites in table/data frame format.}
\usage{
ais_site_config
}
\description{
AIS site configuration and metadata, culled from internal documentation. Lists site
name, domain, type, and data product availabilty.
}
\keyword{AIS,}
\keyword{configuration}
\keyword{metadata,}
