% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2_dq_test.R
\name{co2.dq.test}
\alias{co2.dq.test}
\title{Perform A Commissioning DQ Test on CO2 Concentration Data
#' @author Robert Lee \email{rlee@battelleecology.org}\cr}
\usage{
co2.dq.test(site = site, bgn.month, end.month, save.dir)
}
\arguments{
\item{\code{site}}{Parameter of class character. The NEON site data should be downloaded for.}

\item{\code{bgn.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the first
month to get data for.}

\item{\code{end.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the last
month to get data for.}

\item{\code{save.dir}}{Parameter of class character. The local directory where data files should
be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, variables, and data product or name of family of data
products, data are downloaded and saved to the specifed directory. Process quality calculations are
 then performed and written to a results file in save.dir.
}
\examples{
site = "BART"
bgn.month = "2017-09"
end.month = "2017-10"
save.dir<-"~/Desktop/"
Noble::tis.dq.test(site = site, bgn.month = bgn.month, end.month = end.month, save.dir=save.dir)
}
\seealso{
Currently none
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
