% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{data.pull}
\alias{data.pull}
\title{Downloads data for a specified data product or products, and saves the data to a specified directory}
\usage{
data.pull(site = "JORN", dpID = "DP1.00001.001", bgn.month = "2017-02",
  end.month = "2017-04", time.agr = 30, package = "basic", save.dir)
}
\arguments{
\item{\code{site}}{Parameter of class character. The NEON site data should be downloaded for.}

\item{\code{dpID}}{Parameter of class character. The name of the data product to pull data, or a keyword for a family of data products, e.g. "wind" will pull for 2D and 3D wind data products.}

\item{\code{bgn.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{\code{end.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{\code{time.agr}}{Parameter of class numeric. The data agregation interval requested, must be 1, 2, or 30.}

\item{\code{package}}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{\code{save.dir}}{Parameter of class character. The local directory where data files should be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, package parameters, and data product or name of family of data products,
data are downloaded and saved to the specifed directory.
}
\examples{
#Make a temporary direcotry for the example:
tempDir<- tempdir()
data.pull(site = "CPER", dpID = "Radiation", bgn.month = "2017-02", end.month = "2017-03", time.agr = 30, package="expanded", save.dir = tempDir)
data.pull(site = "CPER", dpID = "DP1.00002.001", bgn.month = "2017-04", end.month = "2017-05", time.agr = 30, package="basic", save.dir= tempDir)
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
