% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_extract.R
\name{date.extract}
\alias{date.extract}
\title{Date Extract}
\usage{
date.extract(data, bgn.date, end.date)
}
\arguments{
\item{data}{A data frame generated by the Noble package.}

\item{bgn.date}{The beginning time to extract data for, as YYYY-MM-DD hh:mm:ss}

\item{end.date}{The end time to extract data for, as YYYY-MM-DD hh:mm:ss}
}
\value{
A data frame with data collected between \code{bgn.time} and \code{end.time}.
}
\description{
Extract data between specified start and end times.
}
\examples{
\dontrun{
data=Noble::pull.data(site = "CPER", dp.id = "DP1.00001.001", bgn.month = "2017-08",
end.month = "2017-08", time.agr = 30, save.dir = tempdir())
## Extract data from Aug 8th, UTC.
aug8=date.extract(data=data, bgn.date = "2017-08-08", end.date = "2017-08-09")
}

}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{data,}
\keyword{date,}
\keyword{subset,}
