% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distorted_field.R
\name{distorted.field}
\alias{distorted.field}
\title{Distorted Field Calculations}
\usage{
distorted.field(site)
}
\arguments{
\item{site}{The NEON TIS site of interest}
}
\value{
A list of distorted field mins and maxes, as well as the corresponding upper and lower
buffer field mins and maxes.
}
\description{
Produces the boundaries of the distorted flow fields for 2D Wind Speed and Direction at
TIS sites
}
\examples{
\dontrun{
out=distorted.field("CPER")
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{2D}
\keyword{data}
\keyword{distorted}
\keyword{flow,}
\keyword{quality}
\keyword{wind,}
