% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_test.R
\name{fan.test}
\alias{fan.test}
\title{Tests fan aspiration system performance}
\usage{
fan.test(site = site, bgn.month, end.month, save.dir, pass.th = 95)
}
\arguments{
\item{site}{Parameter of class character. The NEON site data should be downloaded for.}

\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first
month to get data for.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last
month to get data for.}

\item{save.dir}{Parameter of class character. The local directory where data files should be
saved.}

\item{pass.th}{Optional, defaults to 95. The percent of data that must be unflagged for the
site to pass the test.}
}
\value{
Writes results file to the specified directory.
}
\description{
For the specified dates and site, will test and report results in a results.csv file.
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
