% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_find.R
\name{gap.find}
\alias{gap.find}
\title{Returns indicies for missing NEON data}
\usage{
gap.find(data, time.agr, return)
}
\arguments{
\item{data}{A data frame of NEON data returned from the API or SOM tool.}

\item{time.agr}{Optional, but recommended. The data aggregation period of the input data
(difference between timestamps in minutes), if not specified it will guess at the value.}

\item{return}{Optional. Used to specifiy whether row numbers ("index") or timestamps
("times") are returned. Defaults to row numbers if not specified.}
}
\value{
Outputs a list of lists of row numbers or timestamps where data is missing.
}
\description{
Given a data frame of NEON data, \code{gap.find} will return either a list rows or
list of timestamps. Each returned list is list of missing timestamps, missing data (all NAs),
and data with no values, but that are still quality flagged.\cr

Whether rows of missing data or timestamps are returned depends on the return parameter.
If return = "index", the returned list components will be as follows:\cr
\cr
    miss.indx - The row numbers where timestamps and data were missing\cr
    no.data.indx - The row numbers where data are missing, but are quality flagged\cr
    no.qf.indx - The row numbers where all data and QFs were 'NA'\cr

If return = "times", the returned list components will be as follows:\cr
\cr
    miss.times - The times where timestamps and data were missing\cr
    na.data.times - The times where all data and QFs were 'NA'\cr
    no.qf.times - The times where data are missing, but are quality flagged\cr
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
