% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHCNDstations}
\alias{GHCNDstations}
\title{GHCND Site List}
\format{A data frame/table listing site metadata for 242 USCRN sites.}
\usage{
GHCNDstations
}
\description{
A list of GHCND stations quieriable with \code{Noble::pull.USCRN.data()}. Parsed from
ghcnd-stations.txt, available on the NCDC/NCEI's FTP server at
\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}.
}
\keyword{GHCND,}
\keyword{NCDC,}
\keyword{NCEI,}
\keyword{USCRN,}
\keyword{metadata}
