% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_to_df.R
\name{hdf5.to.df}
\alias{hdf5.to.df}
\title{Convert NEON Eddy Covaraince Data From hdf5 to Data Frames}
\usage{
hdf5.to.df(site, hdf5.file, meas.name, time.agr, save.dir)
}
\arguments{
\item{site}{Parameter of class character. The 4-letter NEON site code that the data is for.}

\item{hdf5.file}{Parameter of class character. The path to the hdf5 file to convert.}

\item{meas.name}{Parameter of class character. The name of the measurement in the hdf5 file
to be extracted.}

\item{time.agr}{What the time difference between sequence values should be, in minutes.}

\item{save.dir}{Optional. If specified a CSV of the extracted data will be saved to the
input directory.}
}
\value{
A data table of mesurements for the requested data product.
}
\description{
This function will extract a given dataset ('meas.name') from the nested hdf5
data structure, and convert it to a data frame. If a save location is specified, a csv of the
data will also be saved.
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{covariance,}
\keyword{data}
\keyword{eddy}
\keyword{gaps,}
\keyword{hdf5,}
\keyword{process}
\keyword{quality,}
