% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_data.R
\name{health.data}
\alias{health.data}
\title{Produce a Summary Table of Data Product Health by Month}
\usage{
health.data(site, dp.id, bgn.month, end.month, save.dir)
}
\arguments{
\item{site}{Parameter of class character. The NEON site of interest.}

\item{dp.id}{Parameter of class character. The NEON data product code of the data product of
interest.}

\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{save.dir}{The directory for data files to be saved to.}
}
\value{
A data frame of health statisitcs by month for a given site. Raw NEON data are also saved to the specified save.dir, if supplied.
}
\description{
For a specified data product ID, this function will produce a data frame of data
product availability and validity for their period of record at a site.


Because the full period of record for all sites are queried,
this function can take a long time to execute.
}
\examples{
# Summarize 2D wind perfomance at CPER:
CPER_wind=dp.survey(dp.id = "DP1.00001.001", site="CPER")
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning,}
\keyword{data}
\keyword{gaps,}
\keyword{health}
\keyword{process}
\keyword{product,}
\keyword{quality,}
