% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{is_site_config}
\alias{is_site_config}
\title{Site Configuration for NEON IS Stations}
\format{Information on all 81 IS sites in table/data frame format.}
\usage{
is_site_config
}
\description{
A merged metadata set of TIS and AIS sites, culled from internal documentation.
Lists site name, domain, type, and data product availabilty. Information in the ais_site_config and
tis_site_config databases was discarded to merge the two databases, so it's recommended that the
AIS- or TIS-specific database be used instead where possible.
}
\keyword{AIS,}
\keyword{TIS,}
\keyword{configuration}
\keyword{metadata,}
