% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.chapter.5.R
\name{make.chapter.5}
\alias{make.chapter.5}
\title{Make Chapter 5 Summary RMD}
\usage{
make.chapter.5(parsed.results)
}
\arguments{
\item{\code{parsed.results}}{The parsed version of the 'results.csv' in the SCA, generated with
\code{Noble::parse.results}.}
}
\value{
RMD code for Chapter 5 of TIS CTRs.
}
\description{
For a data frame of unique site results ('parsed results'), RMD code will be generated
for the CTR.
}
\examples{
None

}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
