% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEON_avail.R
\name{neon.avail}
\alias{neon.avail}
\title{Returns a data frame of data product availability by month}
\usage{
neon.avail(dp.id = "DP1.00001.001")
}
\arguments{
\item{dp.id}{Parameter of class character. The NEON data product code of the data product of interest.}
}
\value{
Outputs a data frame of data product availability by month, where 'x' indicates availability.
}
\description{
For a specified data product ID, a data frame of the availabilty of that product
for all NEON instrumented sites is returned. The output of data product availability is best
interpreted with the base \code{View()} function.
}
\examples{
\dontrun{
wind<-neon.avail(dp.id = "DP1.00001.001")
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
