% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dp_survey.R
\name{plot.dp.survey}
\alias{plot.dp.survey}
\title{Save Summary Graphs of Data Product Health by Month}
\usage{
\method{plot}{dp.survey}(dp.id, save.dir, site, pri.var)
}
\arguments{
\item{\code{dp.id}}{Parameter of class character. The NEON data product code of the data product of interest.}

\item{\code{save.dir}}{The directory for data files and output PNGs to be saved to.}

\item{\code{site}}{Optional. Can specify single site or list of sites of interest.}
}
\value{
Outputs a a PDF of plots data on of all measurement levesl, with one PDF per site.
If only one site is specified, the GGPlot2 object for the summary plot is also returned,
for use in automated report writing.
}
\description{
For a specified data product ID, this function will produce summary plots of data
product availability and validity for their period of record at a site. By default, plots for all
sites with data product availability are generated and saved to the specified directory. If
\code{site} is specified, only plots for the site(s) passed to that parameter are generated and saved.



Because the full period of record for all sites are queried,
this function can take a long time to execute.
}
\examples{
# For 2d Wind, save all plots to the current working directory:
dp.survey(dp.id = "DP1.00001.001", save.dir = getwd())
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning,}
\keyword{data}
\keyword{gaps,}
\keyword{health}
\keyword{process}
\keyword{product,}
\keyword{quality,}
