% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_rose.R
\name{plot.wind.rose}
\alias{plot.wind.rose}
\title{Create wind roses for NEON instrumented sites}
\usage{
\method{plot}{wind.rose}(site, bgn.month, end.month, ml, speed.bins,
  dir.bins)
}
\arguments{
\item{\code{site}}{NEON site to produce the wind rose plot.}

\item{\code{bgn.month}}{The start month for wind data to plot.}

\item{\code{end.month}}{The end month for wind data to plot.}

\item{\code{ml}}{Optional. Used to specifiy what measurement level should be plotted.}

\item{\code{speed.bins}}{Optional. The number of bins for wind speed to be plotted in.}

\item{\code{dir.bins}}{Optional. The number of bins for wind directions to be plotted in.}
}
\value{
Outputs a ggplot2 object of the generated wind roses
}
\description{
For a spceified site and time range, produce a wind rose plot. If the "ml" (measurement level)
parameter is specified, a ggplot2 object for that measurement level is produced. Otherwise, a ggplot2 object of a
faceted plot of all available measurement levels is returned.
}
\examples{
CPER<-plot.wind.rose(site="CPER", bgn.month="2017-01", end.month="2017-02", ml=2, speed.bins=10, dir.bins=36)
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
