% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull.data}
\alias{pull.data}
\title{Downloads data for a specified data product or products,
 and saves the data to a specified directory}
\usage{
pull.data(site, dp.id, bgn.month, end.month, time.agr, package = "basic",
  save.dir, complete.times = F)
}
\arguments{
\item{site}{Parameter of class character. The NEON site data should be downloaded for.}

\item{dp.id}{Parameter of class character. The data product code in question. See
\code{Noble::tis_pri_vars} for a selected list of data product names and codes, or
\url{http://data.neonscience.org/data-product-catalog} for a complete list.}

\item{bgn.month}{Parameter of class character.
The year-month (e.g. "2017-01") of the first month to get data for.}

\item{end.month}{Parameter of class character.
The year-month (e.g. "2017-01") of the last month to get data for.}

\item{time.agr}{Parameter of class numeric.
The data agregation interval requested, must be 1, 2, or 30.}

\item{package}{Parameter of class character. Optional.
The type of data package to be returned If not specified, defaults to basic.}

\item{save.dir}{Parameter of class character.
The local directory where data files should be saved.}

\item{complete.times}{Optional. Parameter of class logical.
Should gaps in the endDateTime column be filled? Defaults to FALSE (No gap filling).}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, package parameters,
and data product or name of family of data products,
data are downloaded and saved to the specifed directory.
}
\examples{
\dontrun{
#Make a temporary directory for the example:
tempDir<- tempdir()
pull.data(site = "CPER", dp.id = "DP1.00002.001", bgn.month = "2017-04", end.month = "2017-05",
time.agr = 30, package="basic", save.dir= tempDir)}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
