% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dp_locs.R
\name{pull.dp.locs}
\alias{pull.dp.locs}
\title{Return the Current Sensor Location Information for a Given Data Product}
\usage{
pull.dp.locs(site, dp.id)
}
\arguments{
\item{site}{Parameter of class character. The 4-letter site code for the NEON
site of interest.}

\item{dp.id}{Parameter of class character. The data product ID of interest,
must be in code format, eg. "DP1.00001.001"}
}
\value{
A data frame of location information for sensors used in the data product
generation.
}
\description{
For a given data product ID and site, a data frame of sensor location
information is returned.
}
\examples{
\dontrun{
2d_wind_locs=Noble::pull.dp.locs(site="CPER", dp.id="DP1.00001.001")
}
}
\seealso{
\itemize{
\item\code{\link{neon.avail}}, which returns a data frame of data product availability by site and month.
\item\code{\link{test.sites}}, which produces a list of sites that have a given data product installed.
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{data,}
\keyword{meta}
\keyword{position}
\keyword{sensor,}
\keyword{spatial,}
