% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_eddy_data.R
\name{pull.eddy.data}
\alias{pull.eddy.data}
\title{Download NEON Eddy Covaraince Data}
\usage{
pull.eddy.data(site, bgn.month, end.month, package, save.dir)
}
\arguments{
\item{site}{Parameter of class character. The 4-letter NEON site code that the data is for.}

\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{package}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{save.dir}{Optional. If specified a CSV of the extracted data will be saved to the
input directory.}
}
\value{
HDF5 data files are saved to the specified directory.
}
\description{
This function downloads HDF5 data from the NEON data portal to the specified save
directory. Files are not collated together by month currently, instead they are saved on a site and
month basis.
}
\examples{
\dontrun{
site="CPER"
bgn.month="2017-04"
end.month="2017-11"
package="basic"
save.dir=tempdir()
pull.eddy.data(site, bgn.month, end.month, package, save.dir)
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{covariance,}
\keyword{data}
\keyword{eddy}
\keyword{gaps,}
\keyword{hdf5,}
\keyword{process}
\keyword{quality,}
