% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_USCRN_data.R
\name{pull.USCRN.data}
\alias{pull.USCRN.data}
\title{Pull data from meterologic stations in the USCRN network}
\usage{
pull.USCRN.data(timeScale, stationID, TimeBgn, TimeEnd)
}
\arguments{
\item{\code{site}}{Parameter of class character. The NEON site data should be downloaded for.}

\item{\code{dp.name}}{Parameter of class character. The name of the data product to pull data, or a keyword for a family of data products, e.g. "wind" will pull for 2D and 3D wind data products.}

\item{\code{bgn.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{\code{end.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{\code{time.agr}}{Parameter of class numeric. The data agregation interval requested, must be 1, 2, or 30.}

\item{\code{package}}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{\code{save.dir}}{Parameter of class character. The local directory where data files should be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, package parameters, and data product or name of family of data products,
data are downloaded and saved to the specifed directory.
}
\examples{
#Make a temporary direcotry for the example:
tempDir<- tempdir()
data.pull(site = "CPER", dp.name = "Radiation", bgn.month = "2017-02", end.month = "2017-03", time.agr = 30, package="expanded", save.dir = tempDir)
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
