% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_dq_info}
\alias{rad_dq_info}
\title{'Hard-coded' TIS Site Information for Radiation DQ Commissioning.}
\format{Information for 9 sites tested for Radiation DQ commissioning in table/data frame format.}
\usage{
rad_dq_info
}
\description{
Data used in TIS radiation data quality commissioning, collected in a database for
convenience.
}
\keyword{SCAN,}
\keyword{TIS,}
\keyword{USCRN,}
\keyword{configuration}
\keyword{metadata,}
\keyword{radiation,}
