% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_temp_plot.R
\name{soil.temp.plot}
\alias{soil.temp.plot}
\title{Plot Soil Temperature Profiles at a Given Time}
\usage{
soil.temp.plot(site, date)
}
\arguments{
\item{site}{Parameter of class character. The TIS site of interest.}

\item{date}{Parameter of class character. The date and time of interest, \strong{to the nearest half hour in UTC}. Format should be \code{YYYY-MM-DD HH:MM}.}
}
\value{
A facet plot soil temperature data by depth.
}
\description{
At a given site and date-time, a plot of the soil temperature profile for
all five soil plots is produced.
}
\examples{
\dontrun{plot=Noble::soil.temp.plot(site="CPER", date="2018-02-28 12:30")}
}
\author{
Robert Lee
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
