% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_extract.R
\name{sp.extract}
\alias{sp.extract}
\title{Refine Data from Multiple Soil Plot Locations Down to a Specified Location}
\usage{
sp.extract(data, sp)
}
\arguments{
\item{data}{A data frame of instrumented pull.dataed with \code{pull.data}.}

\item{sp}{Desired soil plot, given as an integer (eg. sp=1, corresponding to the
first soil plot, closest to the tower)}
}
\value{
Returns a data frame of data just from the specified measurement location
}
\description{
For data downloaded by Noble, and a specified soil plot location, data only
from the location are returned as a data frame.
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
