% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tis_pq_loc_test.R
\name{tis.pq.loc.test}
\alias{tis.pq.loc.test}
\title{Downloads and performs process quality checks on NEON data}
\usage{
tis.pq.loc.test(site, dp.id, prin.vars, bgn.month, end.month, save.dir,
  time.agr = 30, package = "basic", q.th, v.th)
}
\arguments{
\item{site}{Parameter of class character. The NEON site data should be downloaded for.}

\item{dp.id}{Parameter of class character. The name of the data product to pull data, or a
keyword for a family of data products, e.g. "wind" will pull for 2D and 3D wind data products.}

\item{prin.vars}{The principle variables to test (variable names, such as 'windSpeed'). Omit the term 'Mean'.}

\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{save.dir}{Parameter of class character. The local directory where data files should be saved.}

\item{time.agr}{Parameter of class numeric. The data agregation interval requested, must be 1, 2, or 30.}

\item{package}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{q.th}{Parameter of class character. Optional. The threshold for data availability for a passing test, defaults to 0.95.}

\item{v.th}{Parameter of class character. Optional. The threshold for data validity for a passing test, defaults to 0.9.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, variables, and data product or name of family of data products,
data are downloaded and saved to the specifed directory. Process quality calculations are then performed and written to a results file in save.dir.
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
