% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tis_pq_test.R
\name{tis.pq.test}
\alias{tis.pq.test}
\title{Downloads and performs process quality checks on NEON data}
\usage{
tis.pq.test(site = "CPER", dpID = "DP1.00001.001", prin.vars,
  bgn.month = "2017-05", end.month = "2017-06", time.agr = 30,
  package = "basic", save.dir, q.th = 95, v.th = 90)
}
\arguments{
\item{\code{site}}{Parameter of class character. The NEON site data should be downloaded for.}

\item{\code{dpID}}{Parameter of class character. The name of the data product to pull data, or a
keyword for a family of data products, e.g. "wind" will pull for 2D and 3D wind data products.}

\item{\code{prin.vars}}{The principle variables to test (variable names, such as 'windSpeed'). Omit the term 'Mean'.}

\item{\code{bgn.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{\code{end.month}}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{\code{time.agr}}{Parameter of class numeric. The data agregation interval requested, must be 1, 2, or 30.}

\item{\code{package}}{Parameter of class character. Optional. The type of data package to be returned If not specified, defaults to basic.}

\item{\code{save.dir}}{Parameter of class character. The local directory where data files should be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates, site, variables, and data product or name of family of data products,
data are downloaded and saved to the specifed directory. Process quality calculations are then performed and written to a results file in save.dir.
}
\examples{
site = "CPER"
dpID = "DP1.00001.001"
prin.vars<-c("windSpeed", "windDir")
bgn.month = "2017-05"
end.month = "2017-06"
time.agr = 30
package="basic"
save.dir<-"/Users/rlee/Dropbox/GitHub/Commissioning-TIS-rhlee12/Tis2DWindPQ_test"
Noble::tis.pq.test(site = site, dpID = dpID, bgn.month = bgn.month, end.month = end.month,
time.agr = time.agr, package=package, save.dir=save.dir, prin.vars=prin.vars)

}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
