% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tis_pri_vars}
\alias{tis_pri_vars}
\title{TIS Primary Variable database}
\format{Information for 16 TIS data products in table/data frame format.}
\usage{
tis_pri_vars
}
\description{
Information on selected data products, such as the data product name, number, primary
variable reported, units, and commissioning info.
}
\keyword{TIS,}
\keyword{configuration}
\keyword{data}
\keyword{metadata,}
\keyword{products,}
\keyword{units,}
