% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tis_site_config}
\alias{tis_site_config}
\title{Site Configuration for NEON AIS Stations}
\format{Information on all 48 TIS sites in table/data frame format.}
\usage{
tis_site_config
}
\description{
TIS site configuration and metadata, culled from internal documentation. Lists site
name, domain, type, lat/long, timezone, and data product availabilty.
}
\keyword{TIS,}
\keyword{configuration}
\keyword{metadata,}
