% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscrn_find.R
\name{uscrn.colocate}
\alias{uscrn.colocate}
\title{Find USCRN Sites Closest to the Specified Site}
\usage{
uscrn.colocate(site)
}
\arguments{
\item{\code{site}}{The NEON IS site of interest.}
}
\value{
A list with two elememnts: the WBAN for the USCRN site found, and the approximate distance
from the specified NEON site.
}
\description{
Get the WBAN for the USCRN site closest to the specified NEON site, as well as the
approximate distance from the NEON site.
}
\examples{
\dontrun{
cper_colocate=uscrn.colocate("CPER")
}
}
\seealso{

}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{USCRN,}
\keyword{WBAN}
\keyword{colocate,}
\keyword{radius,}
\keyword{site,}
