% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USCRN_sites}
\alias{USCRN_sites}
\title{USCRN Site List}
\format{A data frame/table listing site metadata for 242 USCRN sites.}
\usage{
USCRN_sites
}
\description{
A list of USCRN stations. Parsed from
ghcnd-stations.txt, available on the NCDC/NCEI's FTP server at
\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt}.
}
\keyword{GHCND,}
\keyword{NCDC,}
\keyword{NCEI,}
\keyword{USCRN,}
\keyword{metadata}
