% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose_plot.R
\name{wind.rose.plot}
\alias{wind.rose.plot}
\title{Create wind roses for NEON instrumented sites}
\usage{
wind.rose.plot(site, bgn.month, end.month, ml, speed.bins, dir.bins)
}
\arguments{
\item{site}{NEON site to produce the wind rose plot.}

\item{bgn.month}{The start month for wind data to plot.}

\item{end.month}{The end month for wind data to plot.}

\item{ml}{Optional. Used to specifiy what measurement level should be plotted.}

\item{speed.bins}{Optional. The number of bins for wind speed to be plotted in.}

\item{dir.bins}{Optional. The number of bins for wind directions to be plotted in.}
}
\value{
Outputs a ggplot2 object of the generated wind roses
}
\description{
For a spceified site and time range, produce a wind rose plot. If the "ml" (measurement level)
parameter is specified, a ggplot2 object for that measurement level is produced. Otherwise, a ggplot2 object of a
faceted plot of all available measurement levels is returned.
}
\examples{
\dontrun{
CPER<-plot.wind.rose(site="CPER",
bgn.month="2017-01",
end.month="2017-02",
ml=2,
speed.bins=10,
dir.bins=36)
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
