% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_thresholds}
\alias{wind_thresholds}
\title{ATBD Specific Thresholds: 2D Wind Speed and Direction}
\format{Thresholds and tower information for 2D wind at all IS sites in table/data frame format.}
\usage{
wind_thresholds
}
\description{
A theshold file for the 2D Wind Speed and Direction data product (DP1.00001.001), used
by NEON CI to generate flags and quality metrics. Used in Noble to analyze data quality for
2D Wind Speed and Direction.
}
\keyword{2D}
\keyword{Direction,}
\keyword{Speed}
\keyword{TIS,}
\keyword{Wind}
\keyword{and}
\keyword{flag,}
\keyword{metadata,}
\keyword{quality}
\keyword{thresholds,}
