% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temp_dq_test.R
\name{air.temp.dq.test}
\alias{air.temp.dq.test}
\title{Test NEON Air Temperature data for Stability of Variance, Internal Consistency,
and External Consistency.}
\usage{
air.temp.dq.test(site, bgn.month, end.month, save.dir)
}
\arguments{
\item{site}{Parameter of class character. The NEON site data should be downloaded for.}

\item{bgn.month}{Parameter of class character. The year-month (e.g. "2017-01") of the first month to get data for.}

\item{end.month}{Parameter of class character. The year-month (e.g. "2017-01") of the last month to get data for.}

\item{save.dir}{Parameter of class character. The local directory where data files should be saved.}
}
\value{
Writes data files to the specified directory.
}
\description{
For the specified dates and site, the function will perform variance stability testing
 on overnight data (midnight to 4:00 local time), internal consistency checks on adjacent
 measurement locations, and external consistency checks with the closest NRCS or USCRN site.
}
\examples{
\dontrun{
#Make a temporary direcotry for the example:
tempDir= tempdir()
pull.data(site = "CPER",
dp.id = "DP1.00002.001",
bgn.month = "2017-04",
end.month = "2017-05",
time.agr = 30,
package="basic",
save.dir= tempDir)
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
