% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_data.R
\name{health.data}
\alias{health.data}
\title{Produce a Summary Table of Data Product Health by Month}
\usage{
health.data(site, dp.id, bgn.month, end.month, save.dir)
}
\arguments{
\item{\code{dp.id}}{Parameter of class character. The NEON data product code of the data product of
interest.}

\item{\code{site}}{Parameter of class character. The NEON site of interest.}
}
\value{
Outputs a a PDF of plots data on of all measurement levesl, with one PDF per site.
If only one site is specified, the GGPlot2 object for the summary plot is also returned,
for use in automated report writing.
}
\description{
For a specified data product ID, this function will produce a data frame of data
product availability and validity for their period of record at a site.


Because the full period of record for all sites are queried,
this function can take a long time to execute.
}
\examples{
# Summarize 2D wind perfomance at CPER:
CPER_wind=dp.survey(dp.id = "DP1.00001.001", site="CPER")
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning,}
\keyword{data}
\keyword{gaps,}
\keyword{health}
\keyword{process}
\keyword{product,}
\keyword{quality,}
