% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_extract.R
\name{ml.extract}
\alias{ml.extract}
\title{Refine a data frame with multiple measurement levels down to the specified level}
\usage{
ml.extract(data, ml)
}
\arguments{
\item{\code{data}}{A data frame of instrumented data pulled with \code{data.pull}.}

\item{\code{ml}}{Desired measurement level, given as an integer (eg ml=1, corresponding to the first level at a tower)}
}
\value{
Returns a data frame of data just from the specified measurement level
}
\description{
For data downloaded by Noble, and a specified soil plot location, data only
from the location are returned as a data frame.
}
\examples{
Currently none
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
