% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_n_plot.R
\name{pull.n.plot}
\alias{pull.n.plot}
\title{Returns PDF(s) of data for the specified site and data product}
\usage{
pull.n.plot(sites.req, bgn.month, end.month, dpID, save.dir, data.field,
  package)
}
\arguments{
\item{\code{sites.req}}{The site, or character list of sites to return plots of.}

\item{\code{bgn.month}}{The start month to plot data for.}

\item{\code{end.month}}{The end month to plot data for.}

\item{\code{dpID}}{Parameter of class character. The NEON data product code of the data product of interest.}

\item{\code{save.dir}}{The directory for data files and output PDFs to be saved to.}

\item{\code{data.field}}{Optional. The name of the measurement vaiable to plot. Defaults to the 'core' measurement for most products.}
}
\value{
Outputs a a PDF of plots data on of all measurement levesl, with one PDF per site.
}
\description{
For a specified data product ID, a data frame of the availabilty of that product
for all NEON instrumented sites is returned. The output of data product availability is best
interpreted with the base \code{View()} function.
}
\examples{
# for a variable, "test.dir", holding a valid file path:
pull.n.plot(bgn.month = "2017-04", end.month = "2017-05", dpID = "DP1.00001.001", sites.req = "BLAN", save.dir = getwd(), data.field = "windDirMean")
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr

Cove Sturdevant
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
