% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_USCRN_data.R
\name{pull.USCRN.data}
\alias{pull.USCRN.data}
\title{Pull data from meterologic stations in the USCRN network}
\usage{
\method{pull}{USCRN.data}(timeScale, stationID, TimeBgn, TimeEnd, saveDir)
}
\arguments{
\item{\code{stationID}}{The WBAN ID of a valid USCRN site.}

\item{\code{TimeBgn}}{A YYYY-MM-DD hh:mm:ss formated start time, in UTC.}

\item{\code{TimeEnd}}{A YYYY-MM-DD hh:mm:ss formated end time, in UTC.}

\item{\code{timeScale}}{The time scale of measurements to return.
Enter 'monthly', 'daily', 'hourly', or 'subhourly'.}
}
\value{
Data frame of USCRN data.
}
\description{
Provide a valid GHCND or WBAN ID for a site in the USCRN, the desired
timescale of measurements, and requested begin and end times in POSIX format,
and a data frame will be returned to the global environment.
}
\examples{
\dontrun{
timeScale <- "subhourly"
stationID <- "USW00003047"
TimeBgn <- as.POSIXct("2014-04-01 00:00:01", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC")
TimeEnd <- as.POSIXct("2015-02-01 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz="UCT")

grabUSCRN(timeScale = timeScale, TimeBgn = TimeBgn, TimeEnd = TimeEnd, stationID = stationID)
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{USCRN,}
\keyword{commissioning}
\keyword{data}
\keyword{data,}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
