% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sites.R
\name{test.sites}
\alias{test.sites}
\title{Returns a list of sites where a given data product exists in a given time range}
\usage{
test.sites(dpId, bgn.month, end.month)
}
\arguments{
\item{\code{dpId}}{Parameter of class character. The data product ID of interest, must be in code format, eg. "DP1.00001.001"}

\item{\code{bgn.month}}{Parameter of class character. The start month of the sequence of interest.}

\item{\code{end.month}}{Parameter of class character. The (inclusive) end month of the sequence of interest.}
}
\value{
A list of sites where the data product is available between the two input months.
}
\description{
For a given data product ID and the start and end months, a list of sites where the data product is has data is returned.
}
\examples{
2d_wind=test.sites(dpId="DP1.00001.001", )

}
\seealso{
\code{\link{NEON.avail}}, which returns a data frame of data product availability by site and month.
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
