% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_dir_hist.R
\name{wind.dir.hist}
\alias{wind.dir.hist}
\title{Plot 2D Wind Direction Histograms}
\usage{
wind.dir.hist(site, bgn.month, end.month, save.dir, dir.bins = 72)
}
\arguments{
\item{\code{site}}{The 4-letter NEON site code.}

\item{\code{bgn.month}}{The fisrt month of data to use in plotting.}

\item{\code{end.date}}{The last month of data to use in plotting.}

\item{\code{save.dir}}{The save directory for the output plot.}
}
\value{
A PNG of level-by-level wind direction histograms saved to \code{save.dir}.
}
\description{
Produces a PNG of wind direction histograms for all valid levels TIS towers. Bins are
color-coded by final QF state, and the approximate bounds of the distorted flow fields are shown as
vertical lines.
}
\examples{
\dontrun{
plot.wind.dir.hist(site="CPER", bgn.month="2017-05", end.month="2017-05", save.dir=getwd())
}
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{air}
\keyword{commissioning,}
\keyword{consistency,}
\keyword{data}
\keyword{process}
\keyword{quality,}
\keyword{temperature}
